+++ dueca/ChannelReadToken.cxx
@@ -11,28 +11,28 @@
         license         : EUPL-1.2
 */
 
+#include "GenericCallback.hxx"
+#include "UCallbackOrActivity.hxx"
+#include <cstddef>
 #define ChannelReadToken_cxx
 #include "ChannelReadToken.hxx"
-#include <UnifiedChannel.hxx>
-#include <DataClassRegistry.hxx>
 #include "DataSetConverter.hxx"
 #include <ChannelManager.hxx>
-#include <UCClientHandle.hxx>
-#include <debug.h>
 #include <DCOFunctor.hxx>
+#include <DataClassRegistry.hxx>
+#include <UCClientHandle.hxx>
+#include <UnifiedChannel.hxx>
 #include <dassert.h>
+#include <debug.h>
 
 DUECA_NS_START;
 
-ChannelReadToken::ChannelReadToken(const GlobalId& owner,
-                                   const NameSet& channelname,
-                                   const std::string& dataclassname,
-                                   const std::string& entrylabel,
-                                   Channel::EntryTimeAspect time_aspect,
-                                   Channel::EntryArity arity,
-                                   Channel::ReadingMode rmode,
-                                   double requested_span,
-                                   GenericCallback *when_valid) :
+ChannelReadToken::ChannelReadToken(
+  const GlobalId &owner, const NameSet &channelname,
+  const std::string &dataclassname, const std::string &entrylabel,
+  Channel::EntryTimeAspect time_aspect, Channel::EntryArity arity,
+  Channel::ReadingMode rmode, double requested_span,
+  const UCallbackOrActivity &when_valid) :
   GenericToken(owner, channelname, dataclassname),
 
   handle(NULL),
@@ -42,88 +42,35 @@ ChannelReadToken::ChannelReadToken(const GlobalId& owner,
   channel = ChannelManager::single()->findOrCreateChannel(this->getName());
 
   // read token with optional entry match
-  handle = channel->addReadToken(this, dataclassname, entrylabel,
-                                 isSingleEntryOption(arity) ?
-                                 0xfffe : 0xffff,
-                                 time_aspect, rmode,
-                                 when_valid, requested_span);
+  handle =
+    channel->addReadToken(this, dataclassname, entrylabel,
+                          isSingleEntryOption(arity) ? 0xfffe : 0xffff,
+                          time_aspect, rmode, when_valid, requested_span);
 }
 
-ChannelReadToken::ChannelReadToken(const GlobalId& owner,
-                                   const NameSet& channelname,
-                                   const std::string& dataclassname,
-                                   const std::string& entrylabel,
-                                   Channel::EntryTimeAspect time_aspect,
-                                   Channel::EntryArity arity,
-                                   Channel::ReadingMode rmode,
-                                   double requested_span,
-                                   Activity *when_valid) :
-  GenericToken(owner, channelname, dataclassname),
-
-  handle(NULL),
-  arity(arity)
-{
-  // finds or creates the channel. Id request comes later in constructor
-  channel = ChannelManager::single()->findOrCreateChannel(this->getName());
-
-  // read token with optional entry match
-  handle = channel->addReadToken(this, dataclassname, entrylabel,
-                                 isSingleEntryOption(arity) ?
-                                 0xfffe : 0xffff,
-                                 time_aspect, rmode,
-                                 when_valid, requested_span);
-}
-
-
-ChannelReadToken::ChannelReadToken(const GlobalId& owner,
-                                   const NameSet& channelname,
-                                   const std::string& dataclassname,
-                                   entryid_type attach_entry,
-                                   Channel::EntryTimeAspect time_aspect,
-                                   Channel::EntryArity arity,
-                                   Channel::ReadingMode rmode,
-                                   double requested_span,
-                                   GenericCallback *when_valid) :
+ChannelReadToken::ChannelReadToken(
+  const GlobalId &owner, const NameSet &channelname,
+  const std::string &dataclassname, entryid_type attach_entry,
+  Channel::EntryTimeAspect time_aspect, Channel::EntryArity arity,
+  Channel::ReadingMode rmode, double requested_span,
+  const UCallbackOrActivity &when_valid) :
   GenericToken(owner, channelname, dataclassname),
   handle(NULL),
   arity(arity)
 {
   channel = ChannelManager::single()->findOrCreateChannel(this->getName());
-  handle = channel->addReadToken(this, dataclassname, "", attach_entry,
-                                 time_aspect, rmode,
-                                 when_valid, requested_span);
-}
-
-ChannelReadToken::ChannelReadToken(const GlobalId& owner,
-                                   const NameSet& channelname,
-                                   const std::string& dataclassname,
-                                   entryid_type attach_entry,
-                                   Channel::EntryTimeAspect time_aspect,
-                                   Channel::EntryArity arity,
-                                   Channel::ReadingMode rmode,
-                                   double requested_span,
-                                   Activity *when_valid) :
-  GenericToken(owner, channelname, dataclassname),
-  handle(NULL),
-  arity(arity)
-{
-  channel = ChannelManager::single()->findOrCreateChannel(this->getName());
-  handle = channel->addReadToken(this, dataclassname, "", attach_entry,
-                                 time_aspect, rmode,
-                                 when_valid, requested_span);
+  handle =
+    channel->addReadToken(this, dataclassname, "", attach_entry, time_aspect,
+                          rmode, when_valid, requested_span);
 }
 
 // deprecated variant
-ChannelReadToken::ChannelReadToken(const GlobalId& owner,
-                                   const NameSet& channelname,
-                                   const std::string& dataclassname,
-                                   const std::string& entrylabel,
-                                   Channel::EntryTimeAspect time_aspect,
-                                   Channel::EntryArity arity,
-                                   Channel::ReadingMode rmode,
-                                   double requested_span,
-                                   Channel::TransportClass tclass,
-                                   GenericCallback *when_valid) :
+ChannelReadToken::ChannelReadToken(
+  const GlobalId &owner, const NameSet &channelname,
+  const std::string &dataclassname, const std::string &entrylabel,
+  Channel::EntryTimeAspect time_aspect, Channel::EntryArity arity,
+  Channel::ReadingMode rmode, double requested_span,
+  Channel::TransportClass tclass, GenericCallback *when_valid) :
   GenericToken(owner, channelname, dataclassname),
 
   handle(NULL),
@@ -133,89 +80,69 @@ ChannelReadToken::ChannelReadToken(const GlobalId& owner,
   channel = ChannelManager::single()->findOrCreateChannel(this->getName());
 
   // read token with optional entry match
-  handle = channel->addReadToken(this, dataclassname, entrylabel,
-                                 isSingleEntryOption(arity) ?
-                                 0xfffe : 0xffff,
-                                 time_aspect, rmode,
-                                 when_valid, requested_span);
+  handle =
+    channel->addReadToken(this, dataclassname, entrylabel,
+                          isSingleEntryOption(arity) ? 0xfffe : 0xffff,
+                          time_aspect, rmode, when_valid, requested_span);
 }
 
 // deprecated variant
-ChannelReadToken::ChannelReadToken(const GlobalId& owner,
-                                   const NameSet& channelname,
-                                   const std::string& dataclassname,
-                                   entryid_type attach_entry,
-                                   Channel::EntryTimeAspect time_aspect,
-                                   Channel::EntryArity arity,
-                                   Channel::ReadingMode rmode,
-                                   double requested_span,
-                                   Channel::TransportClass tclass,
-                                   GenericCallback *when_valid) :
+ChannelReadToken::ChannelReadToken(
+  const GlobalId &owner, const NameSet &channelname,
+  const std::string &dataclassname, entryid_type attach_entry,
+  Channel::EntryTimeAspect time_aspect, Channel::EntryArity arity,
+  Channel::ReadingMode rmode, double requested_span,
+  Channel::TransportClass tclass, GenericCallback *when_valid) :
   GenericToken(owner, channelname, dataclassname),
   handle(NULL),
   arity(arity)
 {
   channel = ChannelManager::single()->findOrCreateChannel(this->getName());
-  handle = channel->addReadToken(this, dataclassname, "", attach_entry,
-                                 time_aspect, rmode,
-                                 when_valid, requested_span);
+  handle =
+    channel->addReadToken(this, dataclassname, "", attach_entry, time_aspect,
+                          rmode, when_valid, requested_span);
 }
 
-ChannelReadToken::ChannelReadToken(const GlobalId& owner,
-                                   const NameSet& channelname,
-                                   const std::string& dataclassname,
-                                   entryid_type attach_entry,
-                                   Channel::EntryTimeAspect time_aspect,
-                                   Channel::EntryArity arity,
-                                   Channel::ReadingMode rmode,
-                                   GenericCallback *when_valid,
-                                   unsigned requested_depth) :
+ChannelReadToken::ChannelReadToken(
+  const GlobalId &owner, const NameSet &channelname,
+  const std::string &dataclassname, entryid_type attach_entry,
+  Channel::EntryTimeAspect time_aspect, Channel::EntryArity arity,
+  Channel::ReadingMode rmode, const UCallbackOrActivity &when_valid,
+  unsigned requested_depth) :
   GenericToken(owner, channelname, dataclassname),
   handle(NULL),
   arity(arity)
 {
   channel = ChannelManager::single()->findOrCreateChannel(this->getName());
-  handle = channel->addReadToken(this, dataclassname, "", attach_entry,
-                                 time_aspect, rmode,
-                                 when_valid, 0.0, requested_depth);
+  handle =
+    channel->addReadToken(this, dataclassname, "", attach_entry, time_aspect,
+                          rmode, when_valid, 0.0, requested_depth);
 }
 
+ChannelReadToken::~ChannelReadToken() { channel->removeReadToken(handle); }
 
-ChannelReadToken::~ChannelReadToken()
-{
-  channel->removeReadToken(handle);
-}
+void ChannelReadToken::selectFirstEntry() { channel->selectFirstEntry(handle); }
 
-
-void ChannelReadToken::selectFirstEntry()
-{
-  channel->selectFirstEntry(handle);
-}
-
-const GlobalId& ChannelReadToken::getChannelId() const
+const GlobalId &ChannelReadToken::getChannelId() const
 {
   return channel->getId();
 }
 
-void ChannelReadToken::selectNextEntry()
-{
-  channel->getNextEntry(handle);
-}
+void ChannelReadToken::selectNextEntry() { channel->getNextEntry(handle); }
 
 bool ChannelReadToken::isSequential() const
 {
   return handle->entry->isSequential();
 }
 
-bool ChannelReadToken::haveEntry() const
-{ return handle->entry != NULL; }
+bool ChannelReadToken::haveEntry() const { return handle->entry != NULL; }
 
 const entryid_type ChannelReadToken::getEntryId() const
 {
   return handle->entry ? handle->entry->entry->getId() : entry_end;
 }
 
-const std::string& ChannelReadToken::getEntryLabel() const
+const std::string &ChannelReadToken::getEntryLabel() const
 {
   const static std::string nolabel;
   return handle->entry ? handle->entry->entry->getLabel() : nolabel;
@@ -236,10 +163,8 @@ DataTimeSpec ChannelReadToken::getLatestDataTime() const
   return channel->getLatestDataTime(handle);
 }
 
-
-void ChannelReadToken::
-addTarget(const boost::intrusive_ptr<TriggerTarget>& target,
-          unsigned id)
+void ChannelReadToken::addTarget(
+  const boost::intrusive_ptr<TriggerTarget> &target, unsigned id)
 {
   // TODO: extend to remember added and removed targets, and remove these
   // from the handle at destruction
@@ -251,10 +176,9 @@ addTarget(const boost::intrusive_ptr<TriggerTarget>& target,
   }
 }
 
-const void* ChannelReadToken::getAccess(TimeTickType t_request,
-                                        DataTimeSpec& ts_actual,
-                                        GlobalId& origin,
-                                        uint32_t magic)
+const void *ChannelReadToken::getAccess(TimeTickType t_request,
+                                        DataTimeSpec &ts_actual,
+                                        GlobalId &origin, uint32_t magic)
 {
   if (magic != magic_number) {
     throw ChannelWrongDataType(getChannelId(), getTokenHolder());
@@ -262,24 +186,24 @@ const void* ChannelReadToken::getAccess(TimeTickType t_request,
   return channel->getReadAccess(handle, t_request, origin, ts_actual);
 }
 
-void ChannelReadToken::releaseAccess(const void* data_ptr)
+void ChannelReadToken::releaseAccess(const void *data_ptr)
 {
   assert(data_ptr != NULL);
   channel->releaseReadAccess(handle);
 }
 
-void ChannelReadToken::releaseAccessKeepData(const void* data_ptr)
+void ChannelReadToken::releaseAccessKeepData(const void *data_ptr)
 {
   channel->releaseReadAccessKeepData(handle);
 }
 
 ChannelReadToken::AccessResult
-ChannelReadToken::readAndStoreData(AmorphStore& s, TimeTickType& tsprev)
+ChannelReadToken::readAndStoreData(AmorphStore &s, TimeTickType &tsprev)
 {
   GlobalId origin;
   DataTimeSpec ts_actual;
-  const void* data = channel->getReadAccess
-    (handle, MAX_TIMETICK, origin, ts_actual);
+  const void *data =
+    channel->getReadAccess(handle, MAX_TIMETICK, origin, ts_actual);
 
   if (!data) {
     assert(handle->accessed == NULL);
@@ -308,7 +232,7 @@ ChannelReadToken::readAndStoreData(AmorphStore& s, TimeTickType& tsprev)
       return DataSuccess;
     }
   }
-  catch (const AmorphStoreBoundary& e) {
+  catch (const AmorphStoreBoundary &e) {
     // channel->releaseReadAccess(handle); // this is wrong,
     channel->resetReadAccess(handle);
 
@@ -318,17 +242,16 @@ ChannelReadToken::readAndStoreData(AmorphStore& s, TimeTickType& tsprev)
     // handle->entry->read_index = handle->entry->read_index->getPrevious();
     throw(e);
   }
-  //assert(handle->accessed == NULL);
-  //return DataSuccess;
+  // assert(handle->accessed == NULL);
+  // return DataSuccess;
 }
 
-bool
-ChannelReadToken::readAndPack(AmorphStore& s, DataTimeSpec& ts_actual,
-                              const TimeSpec& tsreq)
+bool ChannelReadToken::readAndPack(AmorphStore &s, DataTimeSpec &ts_actual,
+                                   const TimeSpec &tsreq)
 {
   GlobalId origin;
-  const void* data = channel->getReadAccess
-    (handle, tsreq.getValidityEnd(), origin, ts_actual);
+  const void *data =
+    channel->getReadAccess(handle, tsreq.getValidityEnd(), origin, ts_actual);
 
   if (!data) {
     assert(handle->accessed == NULL);
@@ -341,7 +264,7 @@ ChannelReadToken::readAndPack(AmorphStore& s, DataTimeSpec& ts_actual,
     converter->packData(s, data);
     channel->releaseReadAccess(handle);
   }
-  catch (const AmorphStoreBoundary& e) {
+  catch (const AmorphStoreBoundary &e) {
     channel->resetReadAccess(handle);
     handle->accessed = NULL; //->resetDataAccess(handle);
     // revert the read index to previous element? Will this match with
@@ -354,13 +277,12 @@ ChannelReadToken::readAndPack(AmorphStore& s, DataTimeSpec& ts_actual,
   return true;
 }
 
-bool ChannelReadToken::applyFunctor(DCOFunctor* fnct, TimeTickType time)
+bool ChannelReadToken::applyFunctor(DCOFunctor *fnct, TimeTickType time)
 {
   GlobalId origin;
   DataTimeSpec ts_actual;
   bool res = true;
-  const void* data = channel->getReadAccess
-    (handle, time, origin, ts_actual);
+  const void *data = channel->getReadAccess(handle, time, origin, ts_actual);
   if (!data) {
     assert(handle->accessed == NULL);
     return false;
@@ -378,7 +300,6 @@ bool ChannelReadToken::applyFunctor(DCOFunctor* fnct, TimeTickType time)
   return res;
 }
 
-
 unsigned int ChannelReadToken::getNumVisibleSets(const TimeTickType ts) const
 {
   return channel->getNumVisibleSets(handle, ts);
@@ -389,7 +310,8 @@ bool ChannelReadToken::haveVisibleSets(const TimeTickType ts) const
   return channel->haveVisibleSets(handle, ts);
 }
 
-unsigned int ChannelReadToken::getNumVisibleSetsInEntry(const TimeTickType ts) const
+unsigned int
+ChannelReadToken::getNumVisibleSetsInEntry(const TimeTickType ts) const
 {
   return channel->getNumVisibleSetsInEntry(handle, ts);
 }